move the distribution code from utilities to new distribution module
