+++ HARK/tests/test_approxDstns.py
@@ -3,7 +3,7 @@ This file implements unit tests approximate distributions.
 """
 
 # Bring in modules we need
-import HARK.utilities as util
+import HARK.distribution as distribution
 import unittest
 import numpy as np
 
@@ -16,11 +16,11 @@ class testsForDCEGM(unittest.TestCase):
     def test_mu_normal(self):
         for muNormal in self.muNormals:
             for stdNormal in self.stdNormals:
-                w, x = util.approxNormal(40, muNormal)
+                w, x = distribution.approxNormal(40, muNormal)
                 self.assertTrue(sum(w*x)-muNormal<1e-12)
 
     def test_mu_lognormal_from_normal(self):
         for muNormal in self.muNormals:
             for stdNormal in self.stdNormals:
-                w, x = util.approxLognormalGaussHermite(40, muNormal, stdNormal)
-                self.assertTrue(abs(sum(w*x)-util.calcLognormalStyleParsFromNormalPars(muNormal, stdNormal)[0])<1e-12)
+                w, x = distribution.approxLognormalGaussHermite(40, muNormal, stdNormal)
+                self.assertTrue(abs(sum(w*x)-distribution.calcLognormalStyleParsFromNormalPars(muNormal, stdNormal)[0])<1e-12)
